+++ src/pages/useCrossTabs/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { UseCrossTabs } from "../../demos/useCrossTabs";
+
+const UseCrossTabsPage: React.FC = () => <UseCrossTabs />;
+
+export default UseCrossTabsPage;