+++ src/pages/useCrossTabsMessage/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { UseCrossTabsMessage } from "../../demos/useCrossTabsMessage";
+
+const UseCrossTabsMessagePage: React.FC = () => <UseCrossTabsMessage />;
+
+export default UseCrossTabsMessagePage;