+++ src/pages/useDetectMultiTabs/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { UseDetectMultiTabs } from "../../demos/useDetectMultiTabs";
+
+const UseDetectMultiTabsPage: React.FC = () => <UseDetectMultiTabs />;
+
+export default UseDetectMultiTabsPage;