+++ src/pages/useEffect2/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { UseEffect2 } from "../../demos/useEffect2";
+
+const UseEffect2Page: React.FC = () => <UseEffect2 />;
+
+export default UseEffect2Page;