+++ src/pages/useIsFirstRender/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+import { UseIsFirstRender } from "../../demos/useIsFirstRender/index.js";
+
+const UseIsFirstRenderPage: React.FC = () => <UseIsFirstRender />;
+
+export default UseIsFirstRenderPage;